Fixed bugs with json+mysql+sota server
This version of code is functional w.r.t to sending a json file from
client and server extracting the information from json and inserting the
details to MYSQL database. Still the initial registration is yet to be
complete w.r.t to replying and checking if the client is already
registered.