Added checks and responses before and after registration
The system now is able to check if the incoming registration data is
already exist in MYSQL database and respond to client whether the
registration is successful or it is already registered.

Next Step: Introduce statemachine in client side and make it not to send
registration message if it is already registered.