Bug fixes and paritial caching implementation
Fixed several bugs. Key ones are, recv and send buffers are not cleared
as a result previous data are considered, which results in validation
failures. Also used json_decref to avoid memory leaks.

Also, with this release the download will not wait for long times to
decompress original files as the diff files are now cached.