Restructured release table to support multi ecu.
Corrected the previous implementation which pulls out release
information from client info file and scans the releases. This is wrong.
The new implementation does not scan based on client's input, it assumes
that the releases will be organized as make_model->devices->releases in
a directory structure and scans the release files.