Fixed bug in continue from last download feature
The multi-ecu update changes originally copied the diff files from cache
dir to session dir and then created integrated diff tar file. This
resulted in generation of different sha256sum values. As a result if a
session was interrupted and a new session that continues from last part
received will have a different sha256 sum. This is now fixed!!